Normalize option for `model_shadowgram`
An optional argument `normalize` was added to `model_shadowgram` for
returning shadowgrams not normalized to their sum. This was added to
improve visualization of superimposed shadowgrams. Since we are working
on input of arbitrary matrix from numpy array, a hint was added to
`_erosion` error raised when calling over a non-integer
array.